+++ Assets/Game/Scripts/CPU/Views/CPUStore/CPUStoreMediator.cs
@@ -68,7 +68,7 @@ namespace TurboLabz.InstantChess
 		}
 
 		[ListensTo(typeof(UpdatePlayerBucksDisplaySignal))]
-		public void OnUpdatePlayerBucksDisplay(int playerBucks)
+		public void OnUpdatePlayerBucksDisplay(long playerBucks)
 		{
 			view.UpdatePlayerBucks(playerBucks);
 		}
